Rearchitect the code to match the computeProperties() / refreshRendering() lifecycle.
Also, gut the CalendarBase code to be a thin wrapper that forwards settings
to the current view.

The current view is also responsible for querying the store,
whereas before I think the Calendar queried the store and passed the items
to the view.  Not sure how updates to the items were handled.