Simplify all the grid build code and remove dependency on oldRenderData. The first-child and last-child CSS class will be replaced by :first-child and :last-child CSS pseudo-selectors.
Also, put all the "render data" into the top level of ViewBase, so that
Stateful can properly track and notify about changes to those computed
properties.

renderData.items became ViewBase#visibleItems, to differentiate
from ViewBase#items, which is all the items, not just the visible
ones.  In the future though filtering needs to be done by the store.