Upgrade calendar to connect to a dstore instead of dojo/store, and to do the time range filtering in the dstore.
Upgraded dcalendar/StoreMixin to extend delite/StoreMap and thus to use dstore.
Since delite/StoreMap only supports one store argument, both the
tasks and the decorations are now taken from the same store.  Decoration
items must have a type: ”decoration” property.

Currently the code maintains the this.visibleItems[] and
this.visibleDecorationItems[] arrays, keeping them in sync with
this.renderItems[] which contains both normal tasks and
decorations.  Ideally I think the code should be changed so there’s just
this.renderItems[].

Also, the original code had support for in-place editing of new events, where
the event would appear in the view and in this.items[] even though
it hadn’t yet been added to the store.  That code is still there in this checkin,
but likely doesn’t work anymore, partly because of maintaining the two arrays
this.visibleItems[] and this.visibleDecorationItems[], and partly because of
trickiness with supporting the secondarySheet.
I will likely remove that code in a later checkin.

If the Calendar or Views have innerText, it’s now interpreted as calendar data.
Therefore <d-calendar> etc. markup shouldn’t contain anything.

Also, only the calendar.html test has been upgraded; the other tests are now
broken.

Fixes #1.